+++ graylog2-web-interface/src/components/search/SearchExecutionError.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { Col, Panel, Row } from 'react-bootstrap';
 
-import DocumentationLink from 'components/support/DocumentationLink';
-import DocsHelper from 'util/DocsHelper';
+import { ContactUs } from 'components/support';
 
 const SearchExecutionError = React.createClass({
   propTypes: {
@@ -34,32 +33,7 @@ const SearchExecutionError = React.createClass({
           </Col>
         </Row>
 
-        <Row className="content">
-          <Col md={12}>
-            <div className="support-sources">
-              <h2>Need help?</h2>
-              <p>
-                Do not hesitate to consult the Graylog community if your questions are not answered in the{' '}
-                <DocumentationLink page={DocsHelper.PAGES.WELCOME} text="documentation" />.
-              </p>
-
-              <ul>
-                <li>
-                  <i className="fa fa-group" />&nbsp;
-                  <a href="https://www.graylog.org/community-support/" target="_blank">Community support</a>
-                </li>
-                <li>
-                  <i className="fa fa-github-alt" />&nbsp;
-                  <a href="https://github.com/Graylog2/graylog2-server/issues" target="_blank">Issue tracker</a>
-                </li>
-                <li>
-                  <i className="fa fa-heart" />&nbsp;
-                  <a href="https://www.graylog.org/professional-support" target="_blank">Professional support</a>
-                </li>
-              </ul>
-            </div>
-          </Col>
-        </Row>
+        <ContactUs />
       </div>
     );
   },
