+++ graylog2-web-interface/src/components/support/index.jsx
@@ -1,3 +1,4 @@
+export { default as ContactUs } from './ContactUs';
 export { default as DocumentationLink } from './DocumentationLink';
 export { default as SmallSupportLink } from './SmallSupportLink';
 export { default as SupportLink } from './SupportLink';
